completed with javaTechie tutorial and implementation

+++ mongoDBcrud/src/main/java/com/javaTechie/springBoot/mongoDBcrud/repository/BookRepository.java
@@ -3,6 +3,6 @@ package com.javaTechie.springBoot.mongoDBcrud.repository;
 import com.javaTechie.springBoot.mongoDBcrud.entity.Book;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface BookRepository extends MongoRepository<Book,Long> {
+public interface BookRepository extends MongoRepository<Book,Integer> {
 
 }
